refactor(edusharing-mock): create mocks module
